\name{cts}
\alias{cts}
\docType{data}
\alias{cts}
\title{A group of simulated continuous variables}
\description{
    50 observations on each of 100 variables. The data are simulated such that
variables with similar indices are associated with the degree of association 
decaying as variables are further apart (i.e., a correlation or information
matrix with have larger values near the diagonal). Details are given in the vignette.
}
\usage{data(mpmidata); cts}
\format{A matrix with 50 rows and 100 columns}
