\name{disc}
\alias{disc}
\docType{data}
\title{A group of simulated categorical (discrete) variables}
\description{
    50 observations on each of 75 categorical variables. These variables are 
designed to be similar to categorical single nucleotide polymorphism (SNP) data
which have 3 categories (A, H and B where H represents a heterozygous mutation). 
There are no associations between any of the variables. The variables are stored 
as characters. See the vignette for details
}
\usage{data(mpmidata); disc}
\format{A matrix of characters with 50 rows and 75 columns}
